+++ exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/proto/__init__.py
@@ -49,6 +49,11 @@ Usage
     with tracer.start_as_current_span('foo'):
         print('Hello world!')
 
+You can configure the exporter with the following environment variables:
+
+- :envvar:`OTEL_EXPORTER_JAEGER_INSECURE`
+- :envvar:`OTEL_EXPORTER_JAEGER_ENDPOINT`
+- :envvar:`OTEL_EXPORTER_JAEGER_CERTIFICATE`
 API
 ---
 .. _Jaeger: https://www.jaegertracing.io/