+++ exporter/opentelemetry-exporter-jaeger-thrift/src/opentelemetry/exporter/jaeger/thrift/__init__.py
@@ -56,6 +56,15 @@ Usage
     with tracer.start_as_current_span('foo'):
         print('Hello world!')
 
+You can configure the exporter with the following environment variables:
+
+- :envvar:`OTEL_EXPORTER_JAEGER_USER`
+- :envvar:`OTEL_EXPORTER_JAEGER_PASSWORD`
+- :envvar:`OTEL_EXPORTER_JAEGER_ENDPOINT`
+- :envvar:`OTEL_EXPORTER_JAEGER_AGENT_PORT`
+- :envvar:`OTEL_EXPORTER_JAEGER_AGENT_HOST`
+- :envvar:`OTEL_EXPORTER_JAEGER_AGENT_SPLIT_OVERSIZED_BATCHES`
+
 API
 ---
 .. _Jaeger: https://www.jaegertracing.io/